[GBA] Fix misaligned LDRH/LDRSH behavior
Fixes map tile rendering in Dragon Ball Z - The Legacy of Goku

+++ emulators/components/cpu/arm7tdmi/arm7tdmi.h
@@ -195,7 +195,7 @@ private:
 
     // --- Memory access with alignment handling ---
     u8  read8(u32 addr);
-    u16 read16(u32 addr);
+    u32 read16(u32 addr);
     u32 read32(u32 addr);
     void write8(u32 addr, u8 data);
     void write16(u32 addr, u16 data);
